Moved Search Feature Into it's Own Component, and fixed all imports, Page is functioning fine and home will be a navigation for components.
