+++ .env
@@ -4,7 +4,8 @@ REACT_APP_API_KEY=AIzaSyDj7UDHDC3ta2hR8D9w5qx1CZqdfezKEZA
 #Joanavel/Default API KEY
 # REACT_APP_API_KEY=AIzaSyBLR3ihehfi3OXK-UYtucQQBkpnzaaL
 
-
+# API Key -Sarai 
+# REACT_APP_API_KEY=AIzaSyAx__qIKCVf1mnOA-mP9x7YrGdmAtlFPQE
 # use this fetch statement later on
 
 # fetch(
